+++ Buffs/GoldenAmmoReservation.cs
@@ -7,13 +7,14 @@ namespace GoldenPotions.Buffs
 {
     internal class GoldenAmmoReservation : GoldenBuff
     {
-        public override LocalizedText Description => base.Description.WithFormatArgs(GoldenAmmoReservationPotion.NoConsumeChance);
+        public override LocalizedText Description =>
+            base.Description.WithFormatArgs(GoldenAmmoReservationPotion.NoConsumeChance);
+
         public override int OverwriteBuff => BuffID.AmmoReservation;
 
         public override void SafeUpdate(Player player, ref int buffIndex)
         {
-            player.GetModPlayer<GoldenPotionsPlayer>()
-                    .goldenAmmoReservation = true;
+            player.GetModPlayer<GoldenPotionsPlayer>().goldenAmmoReservation = true;
         }
     }
 }