+++ Buffs/GoldenBuff.cs
@@ -17,6 +17,7 @@ namespace GoldenPotions.Buffs
             {
                 player.ClearBuff(OverwriteBuff);
             }
+
             SafeUpdate(player, ref buffIndex);
         }
 
@@ -26,16 +27,17 @@ namespace GoldenPotions.Buffs
             {
                 npc.DelBuff(buffIndex);
             }
+
             SafeUpdate(npc, ref buffIndex);
         }
 
         /// <summary>
-        /// Use instead of <see cref="Update"/>.
+        /// Use instead of <see cref="Update(Player, ref int)"/>.
         /// </summary>
         public virtual void SafeUpdate(Player player, ref int buffIndex) { }
 
         /// <summary>
-        /// Use instead of <see cref="Update"/>.
+        /// Use instead of <see cref="Update(NPC, ref int)"/>.
         /// </summary>
         public virtual void SafeUpdate(NPC npc, ref int buffIndex) { }
     }