+++ Buffs/GoldenIronskin.cs
@@ -7,7 +7,9 @@ namespace GoldenPotions.Buffs
 {
     internal class GoldenIronskin : GoldenBuff
     {
-        public override LocalizedText Description => base.Description.WithFormatArgs(DefenseIncrease);
+        public override LocalizedText Description =>
+            base.Description.WithFormatArgs(DefenseIncrease);
+
         public override int OverwriteBuff => BuffID.Ironskin;
 
         public override void SafeUpdate(Player player, ref int buffIndex)