+++ Buffs/GoldenShine.cs
@@ -6,8 +6,8 @@ namespace GoldenPotions.Buffs
 {
     internal class GoldenShine : GoldenBuff
     {
-        private static readonly Vector3 CenterRGB = new(0.8f, 0.95f, 1f);
-        private static readonly Vector3 CrossRGB = new(1f, 0.9f, 0.7f);
+        private static readonly Vector3 CenterLight = new(0.8f, 0.95f, 1f);
+        private static readonly Vector3 CrossLight = new(1f, 0.9f, 0.7f);
 
         public override int OverwriteBuff => BuffID.Shine;
 
@@ -15,11 +15,11 @@ namespace GoldenPotions.Buffs
         {
             const int offset = 3 * 16;
             var position = new Vector2(player.Center.X, player.Center.Y);
-            Lighting.AddLight(position, CenterRGB);
-            Lighting.AddLight(position - Vector2.UnitY * offset, CrossRGB);
-            Lighting.AddLight(position - Vector2.UnitX * offset, CrossRGB);
-            Lighting.AddLight(position + Vector2.UnitY * offset, CrossRGB);
-            Lighting.AddLight(position + Vector2.UnitX * offset, CrossRGB);
+            Lighting.AddLight(position, CenterLight);
+            Lighting.AddLight(position - Vector2.UnitY * offset, CrossLight);
+            Lighting.AddLight(position - Vector2.UnitX * offset, CrossLight);
+            Lighting.AddLight(position + Vector2.UnitY * offset, CrossLight);
+            Lighting.AddLight(position + Vector2.UnitX * offset, CrossLight);
         }
     }
 }