+++ Rarities/GoldenRarity.cs
@@ -5,7 +5,7 @@ namespace GoldenPotions.Rarities
 {
     internal class GoldenRarity : ModRarity
     {
-        public override Color RarityColor => new Color(255, 249, 181);
+        public override Color RarityColor => new(255, 249, 181);
 
         public override int GetPrefixedRarity(int offset, float valueMult)
         {